4. ZIM-198, Delete the existing folder D2L_measurement_data_Q5 and add folders of gt_Server and host_Clinet folder with respective codes as per discussion. Tested ok.
