+++ kaushal/gt_server_embedded/include/netrx.h
@@ -26,6 +26,8 @@
 #ifndef NETRX_H
 #define NETRX_H
 
+//#include <arpa/inet.h>
+
 #include "packet.h"
 
 // MACROS
@@ -71,6 +73,8 @@ typedef struct netrx {
   int bind_sock_desc;
   int bind_sock_desc_gt;
   int bind_sock_desc_gt_bridge;
+  
+  char *gt_filename;	// csv filename
   
   // Buffer handling variables
   char *ptr_req_buf;  // pointer to buffer where request is received
